% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{create_pkg}
\alias{create_pkg}
\title{Create new R Markdown package}
\usage{
create_pkg(
  name,
  id_netlify = NULL,
  path = ".",
  incl_roxygen2_meta = TRUE,
  incl_reexports = FALSE,
  incl_sysdata_rmd = FALSE,
  incl_data_rmd = FALSE,
  incl_asciicasts_rmd = FALSE,
  incl_pkgdown_config = TRUE,
  incl_ripgrep_config = TRUE,
  incl_ack_config = FALSE,
  incl_air_config = FALSE
)
}
\arguments{
\item{name}{Package name.}

\item{id_netlify}{Netlify site identifier.}

\item{path}{Path to the new package directory.}

\item{incl_roxygen2_meta}{Whether or not to create a \code{man/roxygen/meta.R} file that i.a. stores the metadata for \href{https://roxygen2.r-lib.org/articles/index-crossref.html#family}{roxygen2's \verb{@family} tags}.}

\item{incl_reexports}{Whether or not to create an \code{R/reexports.R} file prefilled with an opinionated set of magrittr and rlang operator exports.}

\item{incl_sysdata_rmd}{Whether or not to create an \code{Rmd/sysdata.nopurl.Rmd} stub file.}

\item{incl_data_rmd}{Whether or not to create an \code{Rmd/data.nopurl.Rmd} stub file.}

\item{incl_asciicasts_rmd}{Whether or not to create an \code{Rmd/asciicasts.nopurl.Rmd} stub file.}

\item{incl_pkgdown_config}{Whether or not to create a minimal \verb{pkgdown/_pkgdown.yml} configuration file.}

\item{incl_ripgrep_config}{Whether or not to create a custom \code{.rgignore} \href{https://github.com/burntsushi/ripgrep}{ripgrep} ignore file.}

\item{incl_ack_config}{Whether or not to create an \href{https://beyondgrep.com/documentation/}{\code{.ackrc} configuration file}.}

\item{incl_air_config}{Whether or not to create an \href{https://posit-dev.github.io/air/configuration.html}{\code{air.toml} configuration file}. Note that air is not
of much use yet for R Markdown packages.}
}
\value{
\code{path}, invisibly.
}
\description{
Populates the directory specified via \code{path} with all the necessary files for a new R Markdown package.

The \code{DESCRIPTION} file is created using
\code{\link[usethis:use_description]{usethis::use_description()}} and all fields except \code{Package}, \code{URL} and \code{BugReports} are sourced from the \href{https://usethis.r-lib.org/reference/use_description.html}{\code{usethis.description} \R option} if defined.
}
