% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{default_exclusions}
\alias{default_exclusions}
\title{pkgpurl's default lintr exclusions}
\usage{
default_exclusions(excl_inst = TRUE, excl_vignettes = TRUE)
}
\arguments{
\item{excl_inst}{Whether or not to exclude all files under \verb{inst/}. A logical scalar.}

\item{excl_vignettes}{Whether or not to exclude all files under \verb{vignettes/}. A logical scalar.}
}
\value{
A named list of \link[lintr:linters]{lintr::linters}.
}
\description{
Returns an opinionated set of files and folders to be excluded from linting, relative to the root of the package directory. To be used with \code{\link[=lint_rmd]{lint_rmd()}},
\code{\link[lintr:lint]{lintr::lint_dir()}} or \code{\link[lintr:lint]{lintr::lint_package()}}.
}
\examples{
pkgpurl::default_exclusions()
}
\seealso{
\code{\link{default_linters}} and \code{\link[=lint_rmd]{lint_rmd()}}
}
