% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\docType{data}
\name{default_linters}
\alias{default_linters}
\title{pkgpurl's default linters}
\format{
A named list of \link[lintr:linters]{lintr::linters}.
}
\usage{
default_linters
}
\description{
Opinionated set of \link[lintr:linters]{lintr::linters}. Built from \code{\link[lintr:linters_with_defaults]{lintr::linters_with_defaults()}} with lots of customizations. See the \href{https://gitlab.com/rpkg.dev/pkgpurl/-/blob/master/Rmd/sysdata.nopurl.Rmd#default_linters}{relevant source code} for details.
}
\examples{
names(pkgpurl::default_linters)
}
\seealso{
\code{\link[=default_exclusions]{default_exclusions()}} and \code{\link[=lint_rmd]{lint_rmd()}}
}
\keyword{datasets}
