% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{gen_pkgdown_ref}
\alias{gen_pkgdown_ref}
\title{Generate pkgdown reference index}
\usage{
gen_pkgdown_ref(rmd)
}
\arguments{
\item{rmd}{The (R) Markdown file content as a character scalar.}
}
\value{
A list.
}
\description{
Generates the \href{https://pkgdown.r-lib.org/reference/build_reference.html#reference-index}{pkgdown reference index} based on the heading hierarchy structure of
a package's main \code{.Rmd} file.
}
\details{
Basically, all of the pkgdown reference index except \code{desc} keys are derived from \code{rmd}'s \href{https://pandoc.org/MANUAL.html#headings}{Markdown \strong{headings}}
and their \strong{hierarchy}. The \code{desc} keys, however, are assembled from the text paragraphs below some specially named headings (see below).

To be able to unambiguously map an \code{.Rmd} file's heading hierarchy to the pkgdown reference index, some rules are necessary. Attention must be paid to the
fact that, while (R) Markdown supports up to six heading levels (corresponding to HTML's \verb{<h1>}–\verb{<h6>} tags), the pkgdown reference index only supports up
to two (\code{title} and \code{subtitle}).

The following rules define how the reference index is generated:
\enumerate{
\item Headings below a heading named \emph{INTERNAL}, \emph{TEMP} or \emph{TEMPORARY} (case-insensitive, but without any inline formatting) are simply ignored when generating
the reference index.
\item Every heading that is a) inline-formatted as \href{https://pandoc.org/MANUAL.html#verbatim}{verbatim} and b) doesn't contain any whitespace characters is
considered to be the name of a help topic (usually the name of a function or dataset) to be included in the reference index. This maps to the \code{contents}
key of the reference index' YAML.
\item Non-help-topic headings above help topic headings are used as reference index (sub)titles as far as hierarchical nesting allows. More precisely,
non-help-topic headings \emph{of the highest two levels} above the help topic heading are used as title and subtitle, the rest of the headings above the help
topic heading is ignored. This maps to the \code{title} and \code{subtitle} keys of the reference index' YAML.
\item Text paragraph(s) below headings named \emph{DESCRIPTION} (case-insensitive, but without any inline formatting), that in turn are hierarchically below a title
or subtitle heading, will be used as descriptions for the (sub)titles. This maps to the \code{desc} key of the reference index' YAML.
}

When fed to \code{gen_pkgdown_ref()}, the following simplified example \code{.Rmd}...\if{html}{\out{<div class="sourceCode rmd">}}\preformatted{---
editor_options:
  markdown:
    wrap: 80
---

# `main_fn`

The fn below is under no title or subtitle in the pkgdown index.

```\{r\}
main_fn <- function(...) \{...\}
```

# First group of fns

## DESCRIPTION

The description text for *First group of fns*.

## `group_wide_fn`

The fn below is only under a title, but no subtitle in the pkgdown index.

```\{r\}
group_wide_fn <- function(...) \{...\}
```

## Some subgroup

The fns below are both under a title and subtitle in the pkgdown index.

### `subgroup_fn_1`

```\{r\}
subgroup_fn_1 <- function(...) \{...\}
```

### `subgroup_fn_2`

```\{r\}
subgroup_fn_2 <- function(...) \{...\}
```

### Sub-subgroup

The above *Sub-subgroup* heading is ignored for the pkgdown index because the
latter only supports up to two heading levels.

The fn below thus moves one hierarchy level up in the pkgdown index.

#### `sub_subgroup_fn`

```\{r\}
sub_subgroup_fn <- function(...) \{...\}
```

## Another subgroup

### DESCRIPTION

The description text for *Another subgroup*.

### `another_subgroup_fn`

```\{r\}
another_subgroup_fn <- function(...) \{...\}
```

# Second group of fns

## Yet another subgroup

### `yet_another_subgroup_fn`

```\{r\}
yet_another_subgroup_fn <- function(...) \{...\}
```
}\if{html}{\out{</div>}}

...yields this pkgdown index YAML:\preformatted{reference:
- contents: main_fn
- title: First group of fns
  desc: The description text for *First group of fns*.
- contents: group_wide_fn
- subtitle: Some subgroup
- contents:
  - subgroup_fn_1
  - subgroup_fn_2
  - sub_subgroup_fn
- subtitle: Another subgroup
  desc: The description text for *Another subgroup*.
- contents: another_subgroup_fn
- title: Second group of fns
- subtitle: Yet another subgroup
- contents: yet_another_subgroup_fn
}
}
\examples{
\dontrun{
pal::gh_text_file(path = "Rmd/pal.Rmd",
                  owner = "salim-b",
                  name = "pal") |>
  pkgpurl::gen_pkgdown_ref() |>
  yaml::as.yaml() |>
  cat()}
}
