% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{main_rmd}
\alias{main_rmd}
\title{Determine a package's main \code{.Rmd} source file}
\usage{
main_rmd(path = ".")
}
\arguments{
\item{path}{The path to the root of the package directory.}
}
\value{
A character vector, of length 1 if a main \code{.Rmd} is found, otherwise of length 0.
}
\description{
Determines which R Markdown file under \verb{Rmd/} in \code{path} is the package's main source file. If multiple \code{.Rmd} files are present, the one whose name matches
the package name is selected.
}
\keyword{internal}
