% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{run_nopurl_rmd}
\alias{run_nopurl_rmd}
\title{Run \verb{*.nopurl.Rmd} files}
\usage{
run_nopurl_rmd(
  path = ".",
  path_rmd = fs::dir_ls(path = fs::path(path, "Rmd"), recurse = TRUE, type = "file", glob
    = "*.nopurl.Rmd"),
  env = NULL,
  document = TRUE,
  build_and_install = TRUE,
  restart_r_session = TRUE,
  quiet = TRUE,
  roclets = NULL,
  args = getOption("devtools.install.args"),
  dependencies = NA,
  upgrade = "never",
  keep_source = getOption("keep.source.pkgs")
)
}
\arguments{
\item{path}{Path to the root of the package directory.}

\item{path_rmd}{Path(s) to the \code{.Rmd} files to be executed. A character vector.}

\item{env}{Environment to evaluate the \code{.Rmd} files in. If \code{NULL}, the \link[=globalenv]{global environment} is used.}

\item{document}{Whether or not to re-build the package documentation after the last \code{.Rmd} file is executed.}

\item{build_and_install}{Whether or not to build and install the package after each \code{.Rmd} file execution.}

\item{restart_r_session}{Whether or not to restart the R session after the last \code{.Rmd} file is executed. Highly recommended if \code{build_and_install = TRUE},
but only possible when R is run within RStudio.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{roclets}{Character vector of roclet names to use with package.
The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}.}

\item{args}{An optional character vector of additional command line
arguments to be passed to \verb{R CMD INSTALL}. This defaults to the
value of the option \code{"devtools.install.args"}.}

\item{dependencies}{What kinds of dependencies to install. Most commonly
one of the following values:
\itemize{
\item \code{NA}: only required (hard) dependencies,
\item \code{TRUE}: required dependencies plus optional and development
dependencies,
\item \code{FALSE}: do not install any dependencies. (You might end up with a
non-working package, and/or the installation might fail.)
See \link[pak]{Package dependency types} for other possible values and more
information about package dependencies.
}}

\item{upgrade}{When \code{FALSE}, the default, pak does the minimum amount
of work to give you the latest version(s) of \code{pkg}. It will only upgrade
dependent packages if \code{pkg}, or one of their dependencies explicitly
require a higher version than what you currently have. It will also
prefer a binary package over to source package, even if the binary
package is older.

When \code{upgrade = TRUE}, pak will ensure that you have the latest
version(s) of \code{pkg} and all their dependencies.}

\item{keep_source}{If \code{TRUE} will keep the srcrefs from an installed
package. This is useful for debugging (especially inside of RStudio or
Positron). Defaults to \code{getOption("keep.source.pkgs") || !build},
since srcrefs are most useful when the package is installed from its
source directory, i.e. when \code{build = FALSE}.}
}
\value{
\code{path_rmd}, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Executes \code{.Rmd} files which are supposed to contain code not included in the \href{https://r-pkgs.org/structure.html#sec-source-package}{source package}, i.e.
usually outsourced to separate \code{.Rmd} files with the \href{https://pkgpurl.rpkg.dev/dev/reference/purl_rmd.html#-rmd-files-excluded-from-purling}{\code{.nopurl} suffix in their filenames}. Those files are typically used to generate package
data.

If an error is encountered saying \verb{internal error -3 in R_decompress1}, restart the R session and run again.
}
\seealso{
Other high-level functions:
\code{\link{lint_rmd}()},
\code{\link{load_pkg}()},
\code{\link{process_pkg}()},
\code{\link{purl_rmd}()}
}
\concept{high_lvl}
