% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{default_exclusions}
\alias{default_exclusions}
\title{pkgpurl's default lintr exclusions}
\usage{
default_exclusions(excl_vignettes = TRUE)
}
\arguments{
\item{excl_vignettes}{Whether or not to exclude all \code{.Rmd} files under \verb{vignettes/}. A logical scalar.}
}
\value{
A named list of \link[lintr:linters]{lintr::linters}.
}
\description{
Opinionated set of files and folders to be excluded from linting, relative to the package path. To be used with \code{\link[=lint_rmd]{lint_rmd()}}, \code{\link[lintr:lint]{lintr::lint_dir()}} or
\code{\link[lintr:lint]{lintr::lint_package()}}.
}
\examples{
names(pkgpurl::default_linters)
}
\seealso{
\code{\link{default_linters}} and \code{\link[=lint_rmd]{lint_rmd()}}
}
