% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\docType{data}
\name{funky_config}
\alias{funky_config}
\title{pkgpurl package configuration data}
\format{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{key}, \code{default_value}, \code{default_value_dynamic}, \code{require}, and \code{description}.
}
\usage{
funky_config
}
\description{
A \link[tibble:tbl_df-class]{tibble} with data about all possible pkgpurl package configuration options. See \code{\link[funky:config_val]{funky::config_val()}} for more information.
}
\examples{
pkgpurl::funky_config
}
\keyword{datasets}
