% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{lint_rmd}
\alias{lint_rmd}
\title{Lint R Markdown package}
\usage{
lint_rmd(
  path = ".",
  linters = default_linters,
  cache = FALSE,
  relative_path = TRUE,
  exclusions = default_exclusions(excl_vignettes = TRUE),
  pattern = "\\\\.[Rr]([Mm][Dd])?$",
  parse_settings = TRUE,
  show_progress = NULL
)
}
\arguments{
\item{path}{Path to the root of the package directory.}

\item{linters}{A named list of linter functions to apply. See \link[lintr]{linters} for a full list of default and available
linters.}

\item{cache}{When logical, toggle caching of lint results. If passed a character string, store the cache in this
directory.}

\item{relative_path}{if \code{TRUE}, file paths are printed using their path relative to the base directory.
If \code{FALSE}, use the full absolute path.}

\item{exclusions}{exclusions for \code{\link[lintr:exclude]{exclude()}}, relative to the package path.}

\item{pattern}{pattern for files, by default it will take files with any of the extensions
.R, .Rmd, .qmd, .Rnw, .Rhtml, .Rrst, .Rtex, .Rtxt allowing for lowercase r (.r, ...).}

\item{parse_settings}{Logical, default \code{TRUE}. Whether to try and parse the \link[lintr:read_settings]{settings}. Otherwise,
the \code{\link[lintr:default_settings]{default_settings()}} are used.}

\item{show_progress}{Logical controlling whether to show linting progress with a simple text
progress bar \emph{via} \code{\link[utils:txtProgressBar]{utils::txtProgressBar()}}. The default behavior is to show progress in
\code{\link[=interactive]{interactive()}} sessions not running a testthat suite.}
}
\value{
An object of class \code{c("lints", "list")}, each element of which is a \code{"list"} object.
}
\description{
This is a convenience wrapper around \code{\link[lintr:lint]{lintr::lint_dir()}} which is tailored to a typical R Markdown package. To use this function, the
\href{https://github.com/jimhester/lintr/#readme}{lintr} package is required.
}
\details{
To avoid unnecessary noise, all the the generated \verb{R/*.gen.R} files as well as R Markdown vignettes under \verb{vignettes/*.Rmd} are excluded from linting.

This function is registered as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio add-in}, allowing RStudio users to assign a \href{https://support.posit.co/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}{custom shortcut} to it and to invoke it from the \href{https://posit.co/blog/rstudio-1-4-a-quick-tour/#command-palette-shortcuts}{command palette}.
}
\seealso{
\code{\link{default_linters}} and \code{\link[=default_exclusions]{default_exclusions()}}

Other high-level functions:
\code{\link{load_pkg}()},
\code{\link{process_pkg}()},
\code{\link{purl_rmd}()},
\code{\link{run_nopurl_rmd}()}
}
\concept{high_lvl}
