% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl-GEN.R
\name{lint_rmd}
\alias{lint_rmd}
\title{Lint R Markdown package}
\usage{
lint_rmd(path = ".")
}
\arguments{
\item{path}{The path to the root of the package directory.}
}
\description{
This is a convenience wrapper around \code{\link[lintr:lint_dir]{lintr::lint_dir()}} which is tailored to a typical R Markdown package. To use this function, the
\href{https://github.com/jimhester/lintr/#readme}{lintr} package must be installed.
}
