% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{lint_rmd}
\alias{lint_rmd}
\title{Lint R Markdown package}
\usage{
lint_rmd(path = ".", exclude_vignettes = TRUE)
}
\arguments{
\item{path}{The path to the root of the package directory.}

\item{exclude_vignettes}{Whether to exclude all \code{.Rmd} files under \verb{vignettes/}. A logical scalar.}
}
\description{
This is a convenience wrapper around \code{\link[lintr:lint_dir]{lintr::lint_dir()}} which is tailored to a typical R Markdown package. To use this function, the
\href{https://github.com/jimhester/lintr/#readme}{lintr} package is required.
}
\details{
To avoid unnecessary noise, all the the generated \verb{R/*.gen.R} files as well as R Markdown vignettes under \verb{vignettes/*.Rmd} are excluded from linting.

This function is also registered as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio add-in}, allowing RStudio users to assign a \href{https://support.rstudio.com/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}{custom shortcut} to it or to invoke it from the \href{https://www.rstudio.com/blog/rstudio-1-4-a-quick-tour/#command-palette--shortcuts}{command palette}.
}
