% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl-package.R
\docType{package}
\name{pkgpurl-package}
\alias{pkgpurl}
\alias{pkgpurl-package}
\title{pkgpurl: Facilitate Authoring R Packages in the R Markdown File
    Format}
\description{
pkgpurl facilitates R package authoring using
    a literate programming approach. The main idea behind this is to write
    all of the R source code in R Markdown files (`Rmd/*.Rmd`), which
    allows the actual code to be freely mixed with explanatory and
    supplementary information in expressive Markdown format. The main
    object of pkgpurl is to provide a standardized way to compile the bare
    `.R` files from the Prose-enhanced and thus more human-oriented `.Rmd`
    files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/salim_b/r/pkgs/pkgpurl}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/pkgpurl/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
