% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{purl_rmd}
\alias{purl_rmd}
\title{Purl \verb{Rmd/*.Rmd} to \verb{R/*.gen.R}}
\usage{
purl_rmd(
  path = ".",
  add_copyright_notice = funky::config_val("add_copyright_notice"),
  add_license_notice = funky::config_val("add_license_notice"),
  gen_pkgdown_ref = funky::config_val("gen_pkgdown_ref"),
  env = parent.frame()
)
}
\arguments{
\item{path}{Path to the root of the package directory.}

\item{add_copyright_notice}{Whether or not to add a \strong{copyright notice} at the beginning of the generated \code{.R} files as recommended by e.g. the \href{https://www.gnu.org/licenses/gpl-howto.html}{GNU licenses}. The notice consists of the name and description of the program and the word \code{Copyright (C)} followed by the release years and the name(s) of the copyright holder(s), or if not specified, the author(s). The year is always the current year. All the other information is extracted from the package's \code{DESCRIPTION} file. A logical scalar. Only applies if \code{path} \link[pal:is_pkg_dir]{is actually an R package directory}.}

\item{add_license_notice}{Whether or not to add a \strong{license notice} at the beginning of the generated \code{.R} files as recommended by e.g. the \href{https://www.gnu.org/licenses/gpl-howto.html}{GNU licenses}. The license is determined from the package's \code{DESCRIPTION} file and currently only the \href{https://spdx.org/licenses/AGPL-3.0-or-later.html}{\code{AGPL-3.0-or-later} license} is supported. A logical scalar. Only applies if \code{path} \link[pal:is_pkg_dir]{is actually an R package directory}.}

\item{gen_pkgdown_ref}{Whether or not to overwrite \href{https://pkgdown.r-lib.org/}{pkgdown}'s \href{https://pkgdown.r-lib.org/reference/build_reference.html#reference-index}{reference index} in the configuration file \verb{_pkgdown.yml} with an auto-generated one based on the main input file as described in \code{\link[=gen_pkgdown_ref]{gen_pkgdown_ref()}}. A logical scalar. Only applies if \code{path} \link[pal:is_pkg_dir]{is actually an R package directory}, \link[pal:is_pkgdown_dir]{pkgdown is set up} and a \link[=main_rmd]{main R Markdown file} exists.}

\item{env}{Environment to evaluate R Markdown inline code expressions in when generating the pkgdown reference index. Only relevant if
\code{gen_pkgdown_ref = TRUE}.}
}
\value{
\code{path}, invisibly.
}
\description{
This function strives to provide a standardized way to convert all relevant \code{.Rmd} files in the \verb{Rmd/} subdirectory to bare \code{.R} files in the \verb{R/}
subdirectory using \code{\link[knitr:knit]{knitr::purl()}}. It is mainly intended for authoring R packages in the \href{https://rmarkdown.rstudio.com/}{R Markdown file format}.
}
\details{
The generated \code{.R} files will be named the same as the \code{.Rmd} files plus the suffix \code{.gen} to indicate the file was auto-generated. So the file
\code{Rmd/foo.Rmd} for example will be converted to \code{R/foo.gen.R}.

The R Markdown file format allows you to intermingle code with related prose in \href{https://bookdown.org/yihui/rmarkdown/markdown-syntax.html}{Markdown syntax}
optimized for human readability. This facilitates (best) practices which are commonly referred to as
\href{https://en.wikipedia.org/wiki/Literate_programming}{\emph{literate programming}}.

In practice, the main advantage of writing R code in R Markdown is that you don't have to rely on
\href{https://cran.r-project.org/doc/manuals/r-release/R-lang.html#Comments}{\verb{#} comments} to explain, annotate or otherwise elaborate on your code. It also
allows you to easily compile your source code to beautifully looking HTML, PDF etc. files using \code{\link[rmarkdown:render]{rmarkdown::render()}}.

This function is registered as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio add-in}, allowing RStudio users to assign a \href{https://support.rstudio.com/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}{custom shortcut} to it and to invoke it from the \href{https://www.rstudio.com/blog/rstudio-1-4-a-quick-tour/#command-palette--shortcuts}{command palette}.
}
\section{Files excluded from purling}{
\code{purl_rmd()} does not generate an \code{.R} file for each and every R Markdown file in the \verb{Rmd/} subdirectory. Two types of \code{.Rmd} files are excluded from
purling:
\enumerate{
\item Files having the suffix \code{.nopurl} in their name, e.g. \code{Rmd/playground.nopurl.Rmd}.
\item Hidden files \href{https://en.wikipedia.org/wiki/Hidden_file_and_hidden_directory#Unix_and_Unix-like_environments}{as per Unix convention} whose names start
with a dot, e.g. \code{Rmd/.playground.Rmd}.
}

The above convention allows for easy exclusion of specific \code{.Rmd} files from purling. A common case for this are scripts that generate \href{https://r-pkgs.org/data.html#sec-data-sysdata}{package-internal data} from raw sources. Such a script could be stored as \code{Rmd/data.nopurl.Rmd}, so that no corresponding file
under \verb{R/*.R} is generated. For the sake of clarity, it's generally advised to prefer the \code{.nopurl} suffix over hiding files.
}

\seealso{
Other high-level functions:
\code{\link{lint_rmd}()},
\code{\link{load_pkg}()},
\code{\link{process_pkg}()},
\code{\link{run_nopurl_rmd}()}
}
\concept{high_lvl}
