% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpurl.gen.R
\name{run_nopurl_rmd}
\alias{run_nopurl_rmd}
\title{Run \verb{*.nopurl.Rmd} files}
\usage{
run_nopurl_rmd(
  path = ".",
  path_rmd = fs::dir_ls(path = fs::path(path, "Rmd"), recurse = TRUE, type = "file", glob
    = "*.nopurl.Rmd"),
  env = NULL,
  document = TRUE,
  build_and_install = TRUE,
  restart_r_session = TRUE,
  quiet = TRUE,
  roclets = NULL,
  args = getOption("devtools.install.args"),
  dependencies = NA,
  upgrade = "never",
  keep_source = getOption("keep.source.pkgs")
)
}
\arguments{
\item{path}{Path to the root of the package directory.}

\item{path_rmd}{Path(s) to the \code{.Rmd} files to be executed. A character vector.}

\item{env}{Environment to evaluate the \code{.Rmd} files in. If \code{NULL}, the \link[=globalenv]{global environment} is used.}

\item{document}{Whether or not to re-build the package documentation after the last \code{.Rmd} file is executed.}

\item{build_and_install}{Whether or not to build and install the package after each \code{.Rmd} file execution.}

\item{restart_r_session}{Whether or not to restart the R session after the last \code{.Rmd} file is executed. Highly recommended if \code{build_and_install = TRUE},
but only possible when R is run within RStudio.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{roclets}{Character vector of roclet names to use with package.
The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}.}

\item{args}{An optional character vector of additional command line
arguments to be passed to \verb{R CMD INSTALL}. This defaults to the
value of the option \code{"devtools.install.args"}.}

\item{dependencies}{Which dependencies do you want to check?
Can be a character vector (selecting from "Depends", "Imports",
"LinkingTo", "Suggests", or "Enhances"), or a logical vector.

\code{TRUE} is shorthand for "Depends", "Imports", "LinkingTo" and
"Suggests". \code{NA} is shorthand for "Depends", "Imports" and "LinkingTo"
and is the default. \code{FALSE} is shorthand for no dependencies (i.e.
just check this package, not its dependencies).

The value "soft" means the same as \code{TRUE}, "hard" means the same as \code{NA}.

You can also specify dependencies from one or more additional fields,
common ones include:
\itemize{
\item Config/Needs/website - for dependencies used in building the pkgdown site.
\item Config/Needs/coverage for dependencies used in calculating test coverage.
}}

\item{upgrade}{Should package dependencies be upgraded? One of "default", "ask", "always", or "never". "default"
respects the value of the \code{R_REMOTES_UPGRADE} environment variable if set,
and falls back to "ask" if unset. "ask" prompts the user for which out of
date packages to upgrade. For non-interactive sessions "ask" is equivalent
to "always". \code{TRUE} and \code{FALSE} are also accepted and correspond to
"always" and "never" respectively.}

\item{keep_source}{If \code{TRUE} will keep the srcrefs from an installed
package. This is useful for debugging (especially inside of RStudio).
It defaults to the option \code{"keep.source.pkgs"}.}
}
\value{
\code{path_rmd}, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Executes \code{.Rmd} files which are supposed to contain code not included in the \href{https://r-pkgs.org/package-structure-state.html#source-package}{source package}, i.e. usually outsourced to separate \code{.Rmd} files with the \href{https://rpkg.dev/pkgpurl/reference/purl_rmd.html#-rmd-files-excluded-from-purling}{\code{.nopurl} suffix in their filenames}. Those files are typically used to generate package data.

If an error is encountered saying \verb{internal error -3 in R_decompress1}, restart the R session and run again.
}
\seealso{
Other high-level functions:
\code{\link{lint_rmd}()},
\code{\link{load_pkg}()},
\code{\link{process_pkg}()},
\code{\link{purl_rmd}()}
}
\concept{high_lvl}
